avoid deleting nodeodm job if path is not found
There was a problem where the nodeodm job was deleted if the workdir
was not found and then the job was created again. To avoid this return a
boolean based on the creation of the orthophoto folder,
this can be improved and use the client to check for existence of
nodeodm objects.